requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
matplotlib>=3.3.2
Pillow>=8.0.0
requests>=2.25.1 